Adding an ability to use a default implementation for sending history.
