+++ src-java/base-topology/base-storm-topology/src/main/java/org/openkilda/wfm/ContextEmittable.java
@@ -1,0 +1,29 @@
+/* Copyright 2023 Telstra Open Source
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package org.openkilda.wfm;
+
+import org.apache.storm.tuple.Tuple;
+import org.apache.storm.tuple.Values;
+
+/**
+ * This interface represents something that has CommandContext and
+ * knows how to emit the payload via a given Storm stream.
+ */
+public interface ContextEmittable {
+    CommandContext getCommandContext();
+
+    void emitWithContext(String stream, Tuple input, Values payload);
+}
