+++ src-java/base-topology/base-storm-topology/src/main/java/org/openkilda/wfm/HistoryUpdateCarrier.java
@@ -1,0 +1,32 @@
+/* Copyright 2019 Telstra Open Source
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+package org.openkilda.wfm;
+
+import org.openkilda.wfm.share.history.model.FlowHistoryHolder;
+
+/**
+ * Defines a base carrier for processing history updates.
+ */
+public interface HistoryUpdateCarrier {
+    /**
+     * Sends main events to history bolt.
+     */
+    void sendHistoryUpdate(FlowHistoryHolder historyHolder);
+
+    default String getHistoryStreamName() {
+        return "HUB_TO_HISTORY_TOPOLOGY_SENDER";
+    }
+}
